backend/requirements.txt
django==4.2.3
pillow==10.0.0
psycopg2==2.9.7
djangorestframework==3.14.0
markdown==3.0.0
django-filter==1.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0
django-environ==0.11.2
sqlparse==0.4.4
asgiref==3.7.2